Avoid using allprojects and subprojects
