+++ core-tests-android/src/main/java/com/powersync/testutils/UserRow.kt
