+++ gradle/libs.versions.toml
@@ -43,7 +43,6 @@ maven-publish = "0.34.0"
 download-plugin = "5.6.0"
 mokkery = "3.0.0"
 kotlinter = "5.2.0"
-keeper = "0.16.1"
 atomicfu = "0.29.0"
 buildKonfig = "0.17.1"
 
@@ -152,7 +151,6 @@ mavenPublishPlugin = { id = "com.vanniktech.maven.publish", version.ref = "maven
 downloadPlugin = { id = "de.undercouch.download", version.ref = "download-plugin" }
 mokkery = { id = "dev.mokkery", version.ref = "mokkery" }
 kotlinter = { id = "org.jmailen.kotlinter", version.ref = "kotlinter" }
-keeper = { id = "com.slack.keeper", version.ref = "keeper" }
 kotlin-android = { id = "org.jetbrains.kotlin.android", version.ref = "kotlin" }
 kotlin-atomicfu = { id = "org.jetbrains.kotlinx.atomicfu", version.ref = "atomicfu" }
 buildKonfig = { id = "com.codingfeline.buildkonfig", version.ref = "buildKonfig" }
